+++ src/main/scala/service/EnterpriseAssemblerService.scala
@@ -2,7 +2,6 @@ package service
 
 import dao.hbase.{HBaseConnectionManager, HBaseDao}
 import dao.parquet.ParquetDAO
-import dao.csv.CsvDAO
 import org.apache.hadoop.hbase.client.Connection
 import spark.SparkSessionManager
 
@@ -12,7 +11,7 @@ import spark.SparkSessionManager
 trait EnterpriseAssemblerService extends HBaseConnectionManager with SparkSessionManager{
 
   def loadFromCsv{
-    withSpark{ implicit SparkSession => CsvDAO.csvToParquet}
+    withSpark{ implicit SparkSession => ParquetDAO.parquetToHFile(SparkSession)}
   }
 
   def loadFromHFile = withHbaseConnection { implicit connection: Connection => HBaseDao.loadHFiles}
